fix: nvm install

+++ nvm/install.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.35.2/install.sh | bash
+if [[ ! -f "$NVM_DIR/nvm.sh" ]]; then
+  echo "  Installing nvm for you."
+
+  curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.35.2/install.sh | bash
+fi
